# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sign-air-discovery",
    "version": "0.0.1"
  },
  "paths": {
    "/v1/discovery/tsps": {
      "post": {
        "tags": [
          "tsp"
        ],
        "summary": "CreateTSP",
        "description": "Create TSP",
        "operationId": "CreateTSP",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TSPIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TSP was created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TSPOut"
                }
              }
            }
          },
          "409": {
            "description": "TSP with this name already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "TSP Type with this name does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tsp"
        ],
        "summary": "GetTSP",
        "description": "Get TSP",
        "operationId": "GetTSP",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "TSP ID"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "TSP Name"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "TSP country"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 2,
              "description": "Page size",
              "default": 50,
              "title": "Size"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "TSP was retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TSPOut"
                  },
                  "title": "Response Gettsp"
                }
              }
            }
          },
          "404": {
            "description": "TSP with this attr does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/discovery/tsps/{tsp_id}": {
      "patch": {
        "tags": [
          "tsp"
        ],
        "summary": "UpdateTSPByID",
        "description": "Update TSP by ID",
        "operationId": "UpdateTSPByID",
        "parameters": [
          {
            "name": "tsp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tsp Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TSPUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TSP was updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TSPOut"
                }
              }
            }
          },
          "404": {
            "description": "TSP with this attr does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tsp"
        ],
        "summary": "DeleteTSPByID",
        "description": "Delete TSP by ID",
        "operationId": "DeleteTSPByID",
        "parameters": [
          {
            "name": "tsp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tsp Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TSP was deleted successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "TSP with this attr does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/discovery/tsps/{tsp_id}/recommendations": {
      "get": {
        "tags": [
          "tsp"
        ],
        "summary": "GetTSPRecommendationsByID",
        "description": "Get TSP Recommendations by ID",
        "operationId": "GetTSPRecommendationsByID",
        "parameters": [
          {
            "name": "tsp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tsp Id"
            }
          },
          {
            "name": "countries",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Countries"
            }
          },
          {
            "name": "tsp_types",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Tsp Types"
            }
          },
          {
            "name": "time_slots",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Time Slots"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TSPs was retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TSPDB"
                  },
                  "title": "Response Gettsprecommendationsbyid"
                }
              }
            }
          },
          "404": {
            "description": "TSP with this attr does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/discovery/tsps/{tsp_id}/countries": {
      "get": {
        "tags": [
          "tsp"
        ],
        "summary": "GetTSPCountriesByID",
        "description": "Get TSP Countries by ID",
        "operationId": "GetTSPCountriesByID",
        "parameters": [
          {
            "name": "tsp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tsp Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TSP's operating countries were retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Gettspcountriesbyid"
                }
              }
            }
          },
          "404": {
            "description": "TSP with this attr does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tsp"
        ],
        "summary": "UpdateTSPCountriesByID",
        "description": "Update Countries of TSP by ID",
        "operationId": "UpdateTSPCountriesByID",
        "parameters": [
          {
            "name": "tsp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tsp Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "title": "Country List"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TSP's Countries were updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Updatetspcountriesbyid"
                }
              }
            }
          },
          "404": {
            "description": "TSP with this attr does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/discovery/goals": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "CreateGoal",
        "description": "Create Goal",
        "operationId": "CreateGoal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Goal was created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalOut"
                }
              }
            }
          },
          "409": {
            "description": "Goal with this name already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "'code' does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "GetGoal",
        "description": "Get Goal",
        "operationId": "GetGoal",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Goal ID"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Goal Name"
            }
          },
          {
            "name": "tsp_provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Goal TSP Provider"
            }
          },
          {
            "name": "tsp_consumer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Goal TSP Consumer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Size"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Goal was retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoalOut"
                  },
                  "title": "Response Getgoal"
                }
              }
            }
          },
          "404": {
            "description": "Goal Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/discovery/goals/{goal_id}": {
      "patch": {
        "tags": [
          "goals"
        ],
        "summary": "UpdateGoalByID",
        "description": "Update Goal by ID",
        "operationId": "UpdateGoalByID",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Goal was updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalOut"
                }
              }
            }
          },
          "404": {
            "description": "Goal Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/discovery/data-requirements/{data_req_code}": {
      "get": {
        "tags": [
          "data_requirements"
        ],
        "summary": "GetDataRequirementByCode",
        "description": "Get Data Requirement by Code",
        "operationId": "GetDataRequirementByCode",
        "parameters": [
          {
            "name": "data_req_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Req Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data Requirement was retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRequirementOut"
                }
              }
            }
          },
          "404": {
            "description": "'code' does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/discovery/data-requirements/validate": {
      "post": {
        "tags": [
          "data_requirements"
        ],
        "summary": "ValidateDataRequirements",
        "description": "Validate List of Data Requirements",
        "operationId": "ValidateDataRequirements",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "title": "Data Req List"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Data Requirements were validated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRequirementsValidationOut"
                }
              }
            }
          },
          "404": {
            "description": "'code' does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GoalIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Goal Name"
          },
          "tsp_provider": {
            "type": "string",
            "title": "Goal TSP Provider"
          },
          "tsp_consumer": {
            "type": "string",
            "title": "Goal TSP Consumer"
          },
          "data_requirements": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codes Data Requirements for Goal"
          }
        },
        "type": "object",
        "required": [
          "name",
          "tsp_provider",
          "tsp_consumer"
        ],
        "title": "GoalIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Message"
      },
      "TSPDB": {
        "properties": {
          "node_id": {
            "type": "integer",
            "title": "TSP Node ID"
          },
          "id": {
            "type": "string",
            "title": "TSP ID"
          },
          "name": {
            "type": "string",
            "title": "TSP Name"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "id",
          "name"
        ],
        "title": "TSPDB"
      },
...
      "TSPIn": {
        "properties": {
          "id": {
            "type": "string",
            "title": "TSP ID"
          },
          "name": {
            "type": "string",
            "title": "TSP Name"
          },
          "countries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "List of TSP Countries"
          },
          "tsp_type": {
            "type": "string",
            "title": "TSP Type"
          },
          "time_slots": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TimeSlotType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "TSP Available Slot",
            "default": [
              "Mornings",
              "Afternoons",
              "Nights"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "countries",
          "tsp_type"
        ],
        "title": "TSPIn"
      },
      "TSPOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "TSP ID"
          },
          "name": {
            "type": "string",
            "title": "TSP Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "TSPOut"
      },
...
      "TSPUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "TSP Name"
          }
        },
        "type": "object",
        "title": "TSPUpdate"
      },
...
      "TimeSlotType": {
        "type": "string",
        "enum": [
          "Mornings",
          "Afternoons",
          "Nights"
        ],
        "title": "TimeSlotType"
      },
...
    }
  }
}